requirements.txt
langchain
faiss-cpu
sentence-transformers
PyPDF2
transformers
